+++ tasks/deploy/2-deployPostLbpStack.ts
@@ -1,13 +1,8 @@
 import * as TAP_TOKEN_DEPLOY_CONFIG from '@tap-token/config';
 import { TAPIOCA_PROJECTS_NAME } from '@tapioca-sdk/api/config';
-import { TapiocaMulticall } from '@typechain/index';
 import { FeeAmount } from '@uniswap/v3-sdk';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import {
-    createEmptyStratYbAsset__task,
-    loadGlobalContract,
-    loadLocalContract,
-} from 'tapioca-sdk';
+import { loadGlobalContract, loadLocalContract } from 'tapioca-sdk';
 import {
     TTapiocaDeployTaskArgs,
     TTapiocaDeployerVmPass,
@@ -30,6 +25,7 @@ import { buildUsdoMarketOracle } from 'tasks/deployBuilds/oracle/buildUsdoMarket
 import { buildWstethUsdOracle } from 'tasks/deployBuilds/oracle/buildWstethUsdOracle';
 import { deployUniPoolAndAddLiquidity } from 'tasks/deployBuilds/postLbp/deployUniPoolAndAddLiquidity';
 import { DEPLOYMENT_NAMES, DEPLOY_CONFIG } from './DEPLOY_CONFIG';
+import { deployPostLbpStack__postDeploy } from './postDeploy/2-postDeploySetup';
 
 /**
  * @notice Called only after tap-token repo `postLbp1` task
@@ -42,8 +38,8 @@ import { DEPLOYMENT_NAMES, DEPLOY_CONFIG } from './DEPLOY_CONFIG';
  *              - sDAI
  * Post deploy: Arb,Eth
  * !!! Requires TAP and WETH tokens to be in the TapiocaMulticall contract (UniV3 pool creation)
- * !!! Requires TAP and WETH tokens to be in the TapiocaMulticall contract (YB deposit)
- *     - Create empty YB strat for TAP and WETH and register them in YB
+ * !!! Requires TAP and WETH tokens to be in the TapiocaMulticall contract (YB deposit) (1e18 for each)
+ *     - Create empty YB strat for TAP and WETH and register them in YB (TODO)
  *     - Deposit YB assets in YB (TODO)
  *     - Set Seer staleness on testnet
  *
@@ -65,105 +61,9 @@ export const deployPostLbpStack__task = async (
             staticSimulation: false, // Can't runs static simulation because constructor will try to call inexistent contract/function
         },
         tapiocaDeployTask,
-        postDeployTask,
+        deployPostLbpStack__postDeploy,
     );
 };
-
-async function postDeployTask(
-    params: TTapiocaDeployerVmPass<{
-        ratioTap: number;
-        ratioWeth: number;
-        amountTap: string;
-        amountWeth: string;
-    }>,
-) {
-    const {
-        hre,
-        VM,
-        tapiocaMulticallAddr,
-        taskArgs,
-        chainInfo,
-        isTestnet,
-        isHostChain,
-        isSideChain,
-    } = params;
-
-    const { tapToken } = loadContracts__generic(hre, taskArgs.tag);
-
-    // Used in Bar Penrose register
-    await createEmptyStratYbAsset__task(
-        {
-            ...taskArgs,
-            token: tapToken.address,
-            deploymentName: DEPLOYMENT_NAMES.TAP_TOKEN_YB_EMPTY_STRAT,
-        },
-        hre,
-    );
-
-    await createEmptyStratYbAsset__task(
-        {
-            ...taskArgs,
-            token: DEPLOY_CONFIG.MISC[chainInfo.chainId]!.WETH!,
-            deploymentName: DEPLOYMENT_NAMES.WETH_YB_EMPTY_STRAT,
-        },
-        hre,
-    );
-
-    // Set staleness on testnet
-    // isTestnet ? 4294967295 : 86400, // CL stale period, 1 day on prod. max uint32 on testnet
-    if (isTestnet) {
-        const contracts = VM.list();
-        const findContract = (name: string) =>
-            contracts.find((e) => e.name === name);
-
-        const chainLinkUtils = await hre.ethers.getContractAt(
-            'ChainlinkUtils',
-            '',
-        );
-
-        if (isHostChain) {
-            const ethSeerCl = findContract(DEPLOYMENT_NAMES.ETH_SEER_CL_ORACLE);
-            const ethUniCl = findContract(DEPLOYMENT_NAMES.ETH_SEER_UNI_ORACLE);
-            const tap = findContract(DEPLOYMENT_NAMES.TAP_ORACLE);
-            const adbTapOption = findContract(
-                DEPLOYMENT_NAMES.ADB_TAP_OPTION_ORACLE,
-            );
-            const tobTapOption = findContract(
-                DEPLOYMENT_NAMES.TOB_TAP_OPTION_ORACLE,
-            );
-            const reth = findContract(
-                DEPLOYMENT_NAMES.RETH_USD_SEER_CL_MULTI_ORACLE,
-            );
-            const wsteth = findContract(
-                DEPLOYMENT_NAMES.WSTETH_USD_SEER_CL_MULTI_ORACLE,
-            );
-
-            const stalenessToSet = [
-                ethSeerCl,
-                ethUniCl,
-                tap,
-                adbTapOption,
-                tobTapOption,
-                reth,
-                wsteth,
-            ];
-            const calls: TapiocaMulticall.CallStruct[] = [];
-            for (const contract of stalenessToSet) {
-                if (contract) {
-                    calls.push({
-                        target: contract.address,
-                        callData: chainLinkUtils.interface.encodeFunctionData(
-                            'changeDefaultStalePeriod',
-                            [4294967295],
-                        ),
-                        allowFailure: false,
-                    });
-                }
-            }
-            await VM.executeMulticall(calls);
-        }
-    }
-}
 
 async function tapiocaDeployTask(
     params: TTapiocaDeployerVmPass<{
@@ -187,7 +87,10 @@ async function tapiocaDeployTask(
     const owner = tapiocaMulticallAddr;
 
     if (isHostChain) {
-        const { tapToken } = loadContracts__generic(hre, tag);
+        const { tapToken } = deployPostLbpStack__task__loadContracts__generic(
+            hre,
+            tag,
+        );
         await deployUniPoolAndAddLiquidity({
             ...params,
             taskArgs: {
@@ -211,7 +114,8 @@ async function tapiocaDeployTask(
     if (isHostChain) {
         // TapWethLp is used in the oracles, so it must be deployed first
         // Deployment happens above in `deployUniPoolAndAddLiquidity`
-        const { tapToken, tapWethLp } = loadContracts__arb(hre, tag);
+        const { tapToken, tapWethLp } =
+            deployPostLbpStack__task__loadContracts__arb(hre, tag);
 
         VM.add(await buildETHCLOracle(hre, owner, isTestnet))
             .add(await buildETHUniOracle(hre, owner, isTestnet))
@@ -313,7 +217,10 @@ async function tapiocaDeployTask(
     }
 }
 
-function loadContracts__generic(hre: HardhatRuntimeEnvironment, tag: string) {
+export function deployPostLbpStack__task__loadContracts__generic(
+    hre: HardhatRuntimeEnvironment,
+    tag: string,
+) {
     const tapToken = loadGlobalContract(
         hre,
         TAPIOCA_PROJECTS_NAME.TapToken,
@@ -325,8 +232,14 @@ function loadContracts__generic(hre: HardhatRuntimeEnvironment, tag: string) {
     return { tapToken };
 }
 
-function loadContracts__arb(hre: HardhatRuntimeEnvironment, tag: string) {
-    const { tapToken } = loadContracts__generic(hre, tag);
+export function deployPostLbpStack__task__loadContracts__arb(
+    hre: HardhatRuntimeEnvironment,
+    tag: string,
+) {
+    const { tapToken } = deployPostLbpStack__task__loadContracts__generic(
+        hre,
+        tag,
+    );
 
     const tapWethLp = loadLocalContract(
         hre,
